requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
web3==6.20.1
python-dotenv==1.0.1
Jinja2==3.1.4